+++ phpstan.php
@@ -7,4 +7,8 @@ $factory = new Airlst\PhpstanConfig\Factory(['src']);
 return $factory
     ->level(9)
     ->withBleedingEdge()
+    ->unusedPublic(
+        methods: false,
+        localMethods: false,
+    )
     ->create();